separate ALL attractions api in MVC
